project
The front-end and back-end are done, but i need to keep working on the connection between them. Initially, i had the client user and the employee user in different tables but i am trying to make a table to hold all the users so i can make the log in  with rest.